requirements.txt
pandas>=1.3.0
numpy>=1.20.0
plotly>=5.3.0
langchain>=0.0.267
langchain_experimental>=0.0.11
langchain_openai>=0.0.2
python-dotenv>=0.19.0
pytesseract>=0.3.8
pdf2image>=1.16.0
docx2txt>=0.8
opencv-python>=4.5.3
Pillow>=8.3.1
flask>=2.0.1
Werkzeug>=2.0.1
streamlit>=1.22.0 
openpyxl